Avoid installing inner cookie in outer poetry env
When we invoke poetry from the tests the commands are executed in the
outer project poetry env and I do not know how to create an inner env.
This means that the outer env gets dirty and we have all sort of side
effects.

With this modification we use the pre-commit bin of the outer project