feat: add poetry version copier variable
Adding a question allows to have a single point of truth for the poetry
version the project utilizes.

Why not a constant?
Using a copier answer allows for easily updating to new copier versions
in a consistent way, while a constant would require manually changing
the version in all the places.